backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.39.0
google-genai==1.46.0
Markdown==3.8.2
PyJWT==2.10.1
python-decouple==3.8
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
